backend-cloud/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-ninja==1.4.3
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
gunicorn==23.0.0
dj-database-url==2.2.0
langchain==0.3.14
langchain-community==0.3.14
requests==2.32.5
cryptography==45.0.6
pydantic==2.11.7

# Security & Performance
argon2-cffi==23.1.0
django-csp==3.8
django-ratelimit==4.1.0
whitenoise==6.6.0

supabase==2.27.0

google-generativeai==0.3.2

reportlab==4.2.5
qrcode==8.0
Pillow==11.3.0

# ML Dependencies
tensorflow==2.15.0
keras==2.15.0
# tf-keras==2.15.0  <-- Removed
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
opencv-python-headless==4.8.1.78

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Linting & Formatting
ruff==0.9.2
